Refine battery values
In my tests, the ESP runs with the batteries at 3,28. it just fails to start up at below VBAT 3,3V. But if the batteries are given a few seconds to recover, it works. And while it is running, we do get quiet a bit of more runtime if we allow going lower. I'll try to improve the brown-out handling in a later commit.