Refactor EFLedClass methods for consistency and clarity
Some can be static. Others do not modify the state of the lass so could be const. I hope. Not a C++ expert, but my static analyzer told me. :P

It also told me that const on the definitions do not do anything, but only on the actual methods. So also cleaned that up.