Decrease touch-delay limits
I have been testing around in some modes and only now really realized that a 'Release' event is not a ShortTouch.

Which is okay that we do it like this. We basically get three modes out of a button: Tap, ShortPress, LongPress.

But when playing around, it felt very sluggish. I thought the ShortPress mode was broken because it felt like a LongPress. :P
I researched online on UX studies and what kind of reaction people expect from Touch-UIs. Usually, a touch->interaction latency should be below 300ms. This is the Tap -> Response delay and not necessarily a good value for 'ShortPress' above them. But it gives a lower bound. I went with 500ms which does feel like a nice balance between 'I can always tap when I want and don't do accidental ShortPress' vs 'I want to do ShortPress and it does not feel so sluggish.'

Try it out. And let's discuss! :-)